sugestoes da revisao aplicadas

+++ app/model/validator/dadosEvento.py
@@ -21,7 +21,7 @@ class ValidarEvento:
             "data fim inscrição": {"type": "datetime", "required": True},
             "local": {"type": "string", "required": True},
             "carga horária": {"type": "integer", "required": True},
-            "valor": {"type": "integer", "required": True},
+            "valor": {"type": "float", "required": True},
             "arte evento": {"type": "string", "required": True},
             "arte qrcode": {"type": "string", "required": False},
         }
